Time: 23 ms (5.90%), Space: 9.5 MB (85.36%) - LeetHub

+++ 137-single-number-ii/137-single-number-ii.cpp
@@ -2,12 +2,12 @@ class Solution {
 public:
     int singleNumber(vector<int>& nums) 
     {
-        long res = 0, n = nums.size();
+        int res = 0, n = nums.size();
         for(int i = 0; i < 32; i++)
         {
-            long x = 1 << i;
+            int x = 1 << i;
             
-            long sum = 0;
+            int sum = 0;
             for(int j = 0; j < n; j++)
             {
                 if(x & nums[j]) sum++;
@@ -16,7 +16,7 @@ public:
             if(sum % 3) res = res | x;
         }
         
-        return (int)res;
+        return res;
     }
 };
